+++ apps/cold-platform-openai/src/utility.ts
@@ -1,0 +1,3 @@
+export function isImage(extension: string) {
+	return ['png', 'jpg', 'gif', 'bmp', 'tiff', 'jpeg', 'heic'].includes(extension.toLowerCase());
+}