+++ libs/react/src/lib/graphql/queries.ts
@@ -205,7 +205,7 @@ export const GET_ALL_PRODUCTS= gql`
       name
     }
   }
-`
+`;
 
 export const CREATE_PRODUCT_MATERIAL = gql`
   mutation CreateProductMaterial($input: ProductMaterialInsertInput!){
@@ -213,7 +213,7 @@ export const CREATE_PRODUCT_MATERIAL = gql`
       id
     }
   }
-`
+`;
 
 export const CREATE_SUPPLIER = gql`
   mutation CreateOrganizationFacility($input: OrganizationFacilityInsertInput!) {
@@ -229,7 +229,21 @@ export const UPDATE_PRODUCT = gql`
       id
     }
   }
-`
+`;
+
+export const GET_ALL_SUSTAINABILITY_ATTRIBUTES_FOR_ORG = gql`
+  query SustainabilityAttributes($organizationId: ID!, $filter: SustainabilityAttributesListFilter) {
+    sustainabilityAttributes(filter: $filter) {
+      id
+      attributeAssurances(filter: { organization: { id: $organizationId } }) {
+        id
+      }
+      level
+      logoUrl
+      name
+    }
+  }
+`;
 
 export const queries: {
   [key: string]: DocumentNode;
@@ -246,6 +260,7 @@ export const queries: {
   GET_ALL_MATERIALS_FOR_ORG: GET_ALL_MATERIALS_FOR_ORG,
   GET_ALL_SCHEMA_ENUMS: GET_ALL_SCHEMA_ENUMS,
   GET_ALL_SUPPLIERS_FOR_ORG: GET_ALL_SUPPLIERS_FOR_ORG,
+  GET_ALL_SUSTAINABILITY_ATTRIBUTES_FOR_ORG: GET_ALL_SUSTAINABILITY_ATTRIBUTES_FOR_ORG,
   CREATE_MATERIAL: CREATE_MATERIAL,
   CREATE_MATERIAL_SUPPLIER: CREATE_MATERIAL_SUPPLIER,
   GET_ALL_PRODUCTS: GET_ALL_PRODUCTS,